% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handlers.R
\name{.anndata_obs_df}
\alias{.anndata_obs_df}
\title{Extract AnnData.obs as an R data.frame with cell IDs in rownames}
\usage{
.anndata_obs_df(obj)
}
\description{
Extract AnnData.obs as an R data.frame with cell IDs in rownames
}
\keyword{internal}
