% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handlers.R
\name{.anndata_obs_names}
\alias{.anndata_obs_names}
\title{AnnData obs_names / var_names with safe fallback}
\usage{
.anndata_obs_names(obj, n = NULL)
}
\description{
AnnData obs_names / var_names with safe fallback
}
\keyword{internal}
