% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handlers.R
\name{.anndata_obsm_array}
\alias{.anndata_obsm_array}
\title{Pull a 2D embedding from AnnData.obsm}
\usage{
.anndata_obsm_array(obj, key)
}
\description{
Pull a 2D embedding from AnnData.obsm
}
\keyword{internal}
