% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handlers.R
\name{.anndata_obsm_keys}
\alias{.anndata_obsm_keys}
\title{List AnnData.obsm keys whose value is at least 2 columns wide}
\usage{
.anndata_obsm_keys(obj)
}
\description{
List AnnData.obsm keys whose value is at least 2 columns wide
}
\keyword{internal}
