% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_palettes.R
\name{.default_expression_colors}
\alias{.default_expression_colors}
\title{Default scaled-expression heatmap colors (high = red, low = black)}
\usage{
.default_expression_colors()
}
\description{
Default scaled-expression heatmap colors (high = red, low = black)
}
\keyword{internal}
