% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_palettes.R
\name{.default_score_colors}
\alias{.default_score_colors}
\title{Default PhenoMapR score bar colors (low, mid, high)}
\usage{
.default_score_colors()
}
\description{
Default PhenoMapR score bar colors (low, mid, high)
}
\keyword{internal}
