% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_analysis.R
\docType{data}
\name{.MIN_CELLS_PER_PHENO_TAIL_FOR_CT_MARKERS}
\alias{.MIN_CELLS_PER_PHENO_TAIL_FOR_CT_MARKERS}
\title{Minimum cells per phenotype tail required for \code{within_cell_type}
cell-type-specific marker contrasts.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
.MIN_CELLS_PER_PHENO_TAIL_FOR_CT_MARKERS
}
\description{
Minimum cells per phenotype tail required for \code{within_cell_type}
cell-type-specific marker contrasts.
}
\keyword{internal}
