% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_palettes.R
\docType{data}
\name{.rdgy11_brewer}
\alias{.rdgy11_brewer}
\title{ColorBrewer diverging \code{RdGy} palette, 11 classes (standard order)}
\format{
An object of class \code{character} of length 11.
}
\usage{
.rdgy11_brewer
}
\description{
ColorBrewer diverging \code{RdGy} palette, 11 classes (standard order)
}
\keyword{internal}
