% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_palettes.R
\name{.scaled_expr_col_fun_from_colors}
\alias{.scaled_expr_col_fun_from_colors}
\title{Scaled-expression color function from resolved palette colors}
\usage{
.scaled_expr_col_fun_from_colors(scale_clip, expr_colors)
}
\description{
Scaled-expression color function from resolved palette colors
}
\keyword{internal}
