% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handlers.R
\name{process_matrix}
\alias{process_matrix}
\title{Process Matrix Input}
\usage{
process_matrix(mat, pseudobulk = FALSE, group_by = NULL, verbose = TRUE)
}
\description{
Process Matrix Input
}
\keyword{internal}
