% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_reference.R
\name{z_scores_binary}
\alias{z_scores_binary}
\title{Z-scores from two-group t-test per gene}
\usage{
z_scores_binary(expr, group, verbose = TRUE)
}
\description{
Z-scores from two-group t-test per gene
}
\keyword{internal}
