% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_reference.R
\name{z_scores_continuous}
\alias{z_scores_continuous}
\title{Z-scores from correlation per gene (Fisher z or test statistic)}
\usage{
z_scores_continuous(expr, y, verbose = TRUE)
}
\description{
Z-scores from correlation per gene (Fisher z or test statistic)
}
\keyword{internal}
