% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_reference.R
\name{z_scores_survival}
\alias{z_scores_survival}
\title{Z-scores from Cox PH per gene}
\usage{
z_scores_survival(expr, time, event, verbose = TRUE)
}
\description{
Z-scores from Cox PH per gene
}
\keyword{internal}
